+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- IexecSDK constructor
 - `wallet.getAddress()` gives the current wallet address
 - `deal.show()` added key `tasks: { [idx]: [taskid] }` to the resolved value
 - dynamic cast and validation of inputs, invalid inputs throw `ValidationError`.